requirements.txt
python-telegram-bot==13.15
pandas
yfinance
alpaca-trade-api
schedule
